crypto: refactor so that it exports all the necessary references
- Components that 'import' crypto will now use references exported
by 'crypto'.
