+++ source/crypto/src/v1/mod.rs
@@ -1,0 +1,15 @@
+mod utils;
+
+pub use k256::{
+    ecdsa::{
+        signature::{Signer, Verifier},
+        Signature, SigningKey, VerifyingKey,
+    },
+    elliptic_curve::sec1::ToEncodedPoint,
+    SecretKey,
+};
+use k256::{elliptic_curve::ecdh::SharedSecret as SSecret, Secp256k1};
+pub use utils::*;
+
+pub type PublicKey = k256::PublicKey;
+pub type SharedSecret = SSecret<Secp256k1>;
