+++ source/p2p_addr/Cargo.toml
@@ -9,7 +9,6 @@ edition = "2018"
 logger = { path = "../logger" }
 crypto = { path = "../crypto" }
 chrono = "0.4"
-k256 = { version = "0.9.6", features = ["ecdh"] }
 rand_core = "0.6.3"
 serde = { version = "1.0.130", features = ["derive"] }
 utils_net = { path = "../utils_net" }
