+++ source/p2p_addr/src/v1/status.rs
@@ -1,11 +1,4 @@
 use chrono::{DateTime, Duration, Local, Utc};
-pub use k256::{
-    ecdh::EphemeralSecret,
-    ecdsa::{
-        signature::{Signer, Verifier},
-        Signature, SigningKey, VerifyingKey,
-    },
-};
 
 #[derive(Debug, Clone)]
 pub enum AddrStatus {
