+++ source/p2p_discovery/Cargo.toml
@@ -21,7 +21,6 @@ utils_net = { path = "../utils_net" }
 tokio = { version = "1.12.0", features = ["full"] }
 tokio-util = { version = "0.7.2", features = ["full"] }
 futures = "0.3.21"
-k256 = { version = "0.9.6", features = ["ecdh"] }
 log = "0.4.0"
 rand = "0.8.4"
 thiserror = "1.0"
