+++ source/p2p_transport/src/v1/transport.rs
@@ -1,10 +1,10 @@
 use crate::Connection;
-use crypto::{Secp256k1, SharedSecret};
+use crypto::SharedSecret;
 use tokio::sync::RwLock;
 
 pub struct Transport {
     pub conn: RwLock<Connection>,
-    pub shared_secret: SharedSecret<Secp256k1>,
+    pub shared_secret: SharedSecret,
 }
 
 impl Transport {}
