+++ source/p2p_transport_ops/Cargo.toml
@@ -24,7 +24,6 @@ futures = "0.3.21"
 tokio = { version = "1.12.0", features = ["full"] }
 bytes = "1"
 atoi = "0.4.0"
-k256 = { version = "0.9.6", features = ["ecdh"] }
 chrono = "0.4"
 utils_net = { path = "../utils_net" }
