+++ source/saksaha/Cargo.toml
@@ -45,7 +45,6 @@ url = { version = "2", features = ["serde"] }
 wasmer = "2.2.1"
 
 [dev-dependencies]
-k256 = { version = "0.9.6", features = ["ecdh"] }
 sha3 = "0.10.1"
 
 [[bin]]
